complex-agents/doheny-surf-desk/pyproject.toml
[project]
name = "doheny-surf-desk"
version = "0.1.0"
description = "Doheny Surf Desk booking agent with multi-agent workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.3.2",
    "livekit-plugins-openai>=1.3.2",
    "livekit-plugins-silero>=1.3.2",
    "livekit-plugins-turn-detector>=1.3.2",
    "livekit-plugins-noise-cancellation~=0.2.5",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]
